requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.49
asyncpg==0.30.0
alembic==1.14.1
python-dotenv==1.0.1
jinja2==3.1.5
python-multipart==0.0.20
httpx==0.28.1
openai==1.59.3
itsdangerous==2.2.0
aiosqlite==0.20.0
